# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movieflix API",
        "description": "API for Movieflix application",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "List all movies",
                "operationId": "findAll",
                "responses": {
                    "200": {
                        "description": "Return a list of movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Update a movie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Failure updating movie"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Delete a movie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted successfully"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "500": {
                        "description": "Failure deleting movie"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "List all movies by genre",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "description": "Genre name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a list of movies by genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failure listing movies by genre"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier for the movie"
                    },
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "string",
                        "description": "Movie genre"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Unique identifier for the genre"
                            },
                            "name": {
                                "type": "string",
                                "description": "Genre name"
                            }
                        }
                    },
                    "language_id": {
                        "type": "string",
                        "description": "Movie language"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "Unique identifier for the language"
                            },
                            "name": {
                                "type": "string",
                                "description": "Language name"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Number of oscars won by the movie"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie release date"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Movie title"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Movie genre"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Movie language"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Number of oscars won by the movie"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Movie release date"
                    }
                }
            }
        }
    }
}